Generate scaffold for posts.

+++ models/post.go
@@ -15,7 +15,7 @@ type Post struct {
 	ID        uuid.UUID `json:"id" db:"id"`
 	Title     string    `json:"title" db:"title"`
 	Body      string    `json:"body" db:"body"`
-	Votes     int64     `json:"votes" db:"votes"`
+	Votes     int       `json:"votes" db:"votes"`
 	CreatedAt time.Time `json:"created_at" db:"created_at"`
 	UpdatedAt time.Time `json:"updated_at" db:"updated_at"`
 }
